requirements.txt
# MovieRAG Core Dependencies
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
numpy>=1.24.0
faiss-cpu>=1.7.0  # Use faiss-gpu for CUDA support
opencv-python>=4.8.0
tqdm>=4.65.0
yt-dlp>=2023.10.13
cinemagoer>=2023.5.1

# Graph Database
neo4j>=5.14.1

# LLM & Generation
groq>=0.4.2
sentence-transformers>=2.2.2
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# UI
gradio>=4.0.0
httpx>=0.25.0
